Size Grandeur

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int a;
 	float f;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(i));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(l));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %lu byte(s)\n", sizeof(c));
+printf("Size of a int: %lu byte(s)\n", sizeof(i));
+printf("Size of a long int: %lu byte(s)\n", sizeof(l));
+printf("Size of a long long int: %lu byte(s)\n", sizeof(a));
+printf("Size of a float: %lu byte(s)\n", sizeof(f));
 return (0);
 }
